requirements.txt
google-genai>=1.0.0
pinecone>=5.3.0
python-dotenv>=1.0.0
rich>=13.0.0
PyMuPDF>=1.20.0
tenacity>=8.0.0
streamlit>=1.30.0